Update Calendar on Props Change.
This is necessary when you're changing the `selectedDate` prop of CalenderPicker programmatically.

For example, in our app after a user selects a date and clicks submit, we reset the form and set the CalendarPicker selectedDate to new Date(). However, the CalendarPicker calendar doesn't actually update.

This commit adds `componentWillReceiveProps`, which watches for new props passed to CalendarPicker and updates based on the new Props.

Works great! :)